# Increase rendering performance