--- a/napari_arboretum/visualisation/vispy_plotter.py
+++ b/napari_arboretum/visualisation/vispy_plotter.py
@@ -19,6 +19,32 @@ class Bounds:
     ymax: float
 
 
+@dataclass
+class TrackSubvisualProxy:
+    pos: np.ndarray
+    width: int = 3
+    color: np.ndarray = np.array([1.0, 1.0, 1.0, 1.0])
+
+    @property
+    def connex(self):
+        connex = [True] * (self.pos.shape[0] - 1) + [False]
+        return connex
+
+    @property
+    def safe_color(self) -> np.ndarray:
+        if self.color.ndim != 2:
+            safe_color = np.repeat([self.color], self.pos.shape[0], axis=0)
+            return safe_color
+        return self.color
+
+
+@dataclass
+class AnnotationSubvisualProxy:
+    pos: np.ndarray
+    text: str
+    color: str = "white"
+
+
 class VisPyPlotter(TreePlotterQWidgetBase):
     """
     Tree plotter using pyqtgraph as the plotting backend.
@@ -38,7 +64,7 @@ class VisPyPlotter(TreePlotterQWidgetBase):
         self.canvas = scene.SceneCanvas(keys=None, size=(300, 1200))
         self.view = self.canvas.central_widget.add_view()
         self.view.camera = scene.PanZoomCamera()
-        self.tree = TreeVisual(parent=None)
+        self.tree = TreeVisualFast(parent=None)
         self.view.add(self.tree)
 
     def get_qwidget(self) -> QWidget:
@@ -104,6 +130,12 @@ class VisPyPlotter(TreePlotterQWidgetBase):
             pos=np.array([[bounds.xmin - padding, time], [bounds.xmax + padding, time]])
         )
 
+    def draw_tree_visual(self) -> None:
+        """
+        Draw the whole tree.
+        """
+        self.tree.draw_tree()
+
 
 class TreeVisual(scene.visuals.Compound):
     """
@@ -139,6 +171,7 @@ class TreeVisual(scene.visuals.Compound):
             Array of shape (n, 4) specifying RGBA values in range [0, 1] along
             the track.
         """
+
         if id is None:
             visual = scene.visuals.Line(pos=pos, color=color, width=3)
         else:
@@ -171,3 +204,114 @@ class TreeVisual(scene.visuals.Compound):
         while self.subvisuals:
             subvisual = self.subvisuals.pop()
             self.remove_subvisual(subvisual)
+
+    def draw_tree(self) -> None:
+        pass
+
+
+class TreeVisualFast(scene.visuals.Compound):
+    """
+    Tree visual that stores branches as sub-visuals.
+    """
+
+    def __init__(self, parent):
+        super().__init__([])
+        self.parent = parent
+        self.unfreeze()
+        # Keep a reference to tracks we add so their colour can be changed later
+        self.tracks = {}
+        self.edges = []
+        self.annotations = []
+
+        subvisuals = [
+            scene.visuals.Line(color="white", width=3),
+            scene.visuals.Text(
+                anchor_x="left",
+                anchor_y="top",
+                rotation=90,
+                font_size=8,
+                color="white",
+            ),
+        ]
+
+        for visual in subvisuals:
+            self.add_subvisual(visual)
+
+    def get_branch_color(self, branch_id: int) -> np.ndarray:
+        return self.tracks[branch_id].color
+
+    def set_branch_color(self, branch_id: int, color: np.ndarray) -> None:
+        """
+        Set the color of an individual branch.
+        """
+        self.tracks[branch_id].color = color
+        self._subvisuals[0].set_data(
+            color=np.row_stack([e.safe_color for e in self.edges]),
+        )
+
+    def add_track(self, id: Optional[int], pos: np.ndarray, color: np.ndarray) -> None:
+        """
+        Parameters
+        ----------
+        id :
+            Track ID.
+        pos :
+            Array of shape (2, 2) specifying vertex coordinates.
+        color :
+            Array of shape (n, 4) specifying RGBA values in range [0, 1] along
+            the track.
+        """
+
+        if id is None:
+            subvisual_proxy = TrackSubvisualProxy(
+                pos=pos,
+                color=np.array([1.0, 1.0, 1.0, 1.0]),
+                width=3,
+            )
+        else:
+            # Split up line into individual time steps so color can vary
+            # along the line
+            ys = np.arange(pos[0, 1], pos[1, 1] + 1)
+            xs = np.ones(ys.size) * pos[0, 0]
+            subvisual_proxy = TrackSubvisualProxy(
+                pos=np.column_stack((xs, ys)),
+                color=color,
+            )
+            # store a reference to this subvisual proxy
+            self.tracks[id] = subvisual_proxy
+
+        self.edges.append(subvisual_proxy)
+
+    def add_annotation(self, x: float, y: float, label: str, color):
+
+        subvisual_proxy = AnnotationSubvisualProxy(
+            text=label,
+            pos=[y, x, 0],
+        )
+
+        self.annotations.append(subvisual_proxy)
+
+    def clear(self) -> None:
+        """Remove all tracks."""
+        self.tracks = {}
+        self.edges = []
+        self.annotations = []
+
+        for visual in self._subvisuals:
+            visual._pos = None
+
+            if hasattr(visual, "_text"):
+                visual._text = None
+
+    def draw_tree(self) -> None:
+        """Once the data is added, draw the tree."""
+
+        self._subvisuals[0].set_data(
+            pos=np.row_stack([e.pos for e in self.edges]),
+            color=np.row_stack([e.safe_color for e in self.edges]),
+            connect=np.concatenate([e.connex for e in self.edges]),
+        )
+
+        # TextVisual does not have a ``set_data`` method
+        self._subvisuals[1].pos = np.asarray([a.pos for a in self.annotations])
+        self._subvisuals[1].text = [a.text for a in self.annotations]