I did some profiling of the rendering code using `examples/show_large_tree.py` and 255 nodes (depth 8):

```sh
255 total nodes
         7509808 function calls (7448213 primitive calls) in 4.881 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.001    0.001    4.881    4.881 base_plotter.py:50(draw_from_nodes)
      892    0.005    0.000    4.370    0.005 visuals.py:125(__init__)
      509    0.002    0.000    2.784    0.005 vispy_plotter.py:110(add_branch)
      892    0.004    0.000    2.626    0.003 visual.py:325(__init__)
      509    0.004    0.000    2.552    0.005 vispy_plotter.py:156(add_track)
```

It looks like a lot of the time is in adding subvisuals, so I had a go at setting things up with a single subvisual for all branches. This gets the performance to a much better state:

```sh
255 total nodes
         121140 function calls (119670 primitive calls) in 0.138 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.140    0.140 base_plotter.py:50(draw_from_nodes)
      509    0.001    0.000    0.100    0.000 vispy_plotter.py:110(add_branch)
      509    0.006    0.000    0.086    0.000 vispy_plotter.py:88(autoscale_view)
6777/6009    0.023    0.000    0.046    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
        1    0.009    0.009    0.036    0.036 tree.py:35(layout_tree)
```

0.14 seconds versus 4.88 seconds.  We will probably want to look at the graph searching algorithms at some point, but this is a good start.

This also merges the small changes in #77, so I will close that one.

![napari-arboretum](https://user-images.githubusercontent.com/8217795/175921897-e6897c4b-b081-4b60-92b5-0e3584433961.png)