@KristinaUlicna 2022-07-04
Looks great! :) All trees show, the single-branch tracks are centred and the performance seems to be better than before (didn't run into the kernel crashing yet). But in the meantime, we lost the y-axis next to the tree [as in #46].